Build C code for prefetches complete fix minor bugs in trans.c

+++ Robust/src/Analysis/Prefetch/IndexDescriptor.java
@@ -57,7 +57,7 @@ public class IndexDescriptor extends Descriptor {
 	}
 
 	public int getOffset() {
-		return offset;
+		return offset.intValue();
 	}
 
 	public String toString() {
